+++ src/Phan/Language/Element/Comment/ReturnComment.php
@@ -3,7 +3,6 @@ declare(strict_types=1);
 
 namespace Phan\Language\Element\Comment;
 
-use Phan\Language\Type\TemplateType;
 use Phan\Language\UnionType;
 
 /**
@@ -46,17 +45,6 @@ class ReturnComment
     public function getLineno() : int
     {
         return $this->lineno;
-    }
-
-    /**
-     * Replace the resolved reference to class T (possibly namespaced) with a regular template type.
-     *
-     * @param array<string,TemplateType> $template_fix_map maps the incorrectly resolved name to the template type
-     * @return void
-     */
-    public function convertTypesToTemplateTypes(array $template_fix_map)
-    {
-        $this->type = $this->type->withConvertTypesToTemplateTypes($template_fix_map);
     }
 
     /**