+++ src/Phan/Plugin/Internal/ThrowAnalyzerPlugin.php
@@ -183,7 +183,7 @@ class ThrowVisitor extends PluginAwarePostAnalysisVisitor
 
     protected static function calculateConfiguredIgnoreThrowsUnionType() : UnionType
     {
-        $throws_union_type = new UnionType();
+        $throws_union_type = UnionType::empty();
         foreach (Config::getValue('exception_classes_with_optional_throws_phpdoc') as $type_string) {
             if (!\is_string($type_string) || $type_string === '') {
                 continue;